Add a server component to the Bikeshed 'Tree' example
git-svn-id: https://google-web-toolkit.googlecode.com/svn/trunk@7607 8db76d5a-ed1c-0410-87a9-c151d255dfc7

+++ bikeshed/src/com/google/gwt/sample/stocks/client/Stocks.java
@@ -316,7 +316,7 @@ public class Stocks implements EntryPoint {
   }
 
   private String getFormattedPrice(int price) {
-    return NumberFormat.getCurrencyFormat("USD").format((double) price / 100.0);
+    return NumberFormat.getCurrencyFormat("USD").format(price / 100.0);
   }
   
   /**
